requirements.txt
# Core Django
Django>=5.2

# Django REST Framework
djangorestframework

# JWT Authentication
djangorestframework-simplejwt

# PostgreSQL
psycopg2

# Environment Variables
python-dotenv

# Image & File Upload
Pillow

# API Documentation (Swagger)
drf-yasg

# CORS (for frontend React / mobile apps)
django-cors-headers

# Celery for async tasks (emails, SMS, notifications)
celery
redis

# Optional — For better password hashing
# argon2-cffi

# Django Filters (required by medical app)
django-filter
